+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Changed
 
 - Rename .master to .ocean with automatic migration at next start
-- Rename ocean workspace -&gt; ocean
+- Rename ocean workspace -> ocean
+- Update version
 
 ## 14.0.5 - 2026-08-04